requirements.txt
flask
WTForms
mysql_connector